format

+++ bin/clearDNS.sh
@@ -1,6 +1,11 @@
-#!/usr/bin/env sh
+#!/bin/sh
 
 ## TODO surround with "if macOS" conditional
-sudo dscacheutil -flushcache
-sudo killall -HUP mDNSResponder
-say cache flushed
+if [ "$(uname -s)" = "Darwin" ]; then # Check if using macOS
+    sudo dscacheutil -flushcache
+    sudo killall -HUP mDNSResponder
+    say cache flushed
+elif [ "$(uname -s)" = "Linux" ]; then
+    sudo /etc/init.d/dns-clean restart
+    sudo /etc/init.d/networking force-reload
+fi
